Install VNDK 28 libprotobuf-cpp-*.so to /vendor
Some vendor prebuilts reference libprotobuf-cpp-lite.so and
libprotobuf-cpp-full.so and expect the 3.0.0-beta3 version.
The new version of protobuf will be installed as
/vendor/lib64/libprotobuf-cpp-lite-3.9.1.so.  The VNDK doesn't
help here because we compile old devices against master and not
an old VNDK snapshot.  We need to continue to provide a vendor
libprotobuf-cpp-lite.so until all products in master get updated
prebuilts or are obsoleted.

Bug: 139175488
Test: m checkbuild
Change-Id: Ida9f89eebd66a11f4cb35857695b425d08d6d40c